+++ monitor/src/settings.rs
@@ -37,6 +37,7 @@ pub struct UiConfig {
 #[derive(Debug, Deserialize)]
 pub struct Settings {
     pub api_endpoint: String,
+    pub job_update_interval: u64,
     pub log_level: Option<u64>,
     pub ui: UiConfig,
     pub rotctld_address: Option<String>,
@@ -49,6 +50,7 @@ impl Settings {
     pub fn new() -> Result<Self, ConfigError> {
         let mut settings = Config::new();
         settings.set_default("api_endpoint", "https://network.satnogs.org/api/")?;
+        settings.set_default("job_update_interval", 600)?;
         settings.set_default("log_level", 0)?;
         settings.set_default("ui.db_min", -100.0)?;
         settings.set_default("ui.db_max", 0)?;